Fix TypeTraits floating point smallest int tests

+++ Source/Test/include/PantheonTest/Tests/TypeTraitsTest.h
@@ -48,11 +48,11 @@ namespace PantheonTest
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<uint16_t>, int16_t>);
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<int32_t>, int32_t>);
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<uint32_t>, int32_t>);
-    static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<float>, int>);
+    static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<float>, int32_t>);
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<int64_t>, int64_t>);
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<uint64_t>, int64_t>);
-    static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<double>, long long>);
-    static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<long double>, long long>);
+    static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<double>, int64_t>);
+    static_assert(std::is_same_v<PantheonCore::Utility::SmallestIntT<long double>, int64_t>);
 
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<int8_t>, uint8_t>);
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<uint8_t>, uint8_t>);
@@ -60,9 +60,9 @@ namespace PantheonTest
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<uint16_t>, uint16_t>);
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<int32_t>, uint32_t>);
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<uint32_t>, uint32_t>);
-    static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<float>, unsigned int>);
+    static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<float>, uint32_t>);
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<int64_t>, uint64_t>);
     static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<uint64_t>, uint64_t>);
-    static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<double>, unsigned long long>);
-    static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<long double>, unsigned long long>);
+    static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<double>, uint64_t>);
+    static_assert(std::is_same_v<PantheonCore::Utility::SmallestUIntT<long double>, uint64_t>);
 }
